installed node js, gulp, jasmine+ first teststory
The reason I installed these is to practice with node js and gulp and to
find out how jasmine npm works, thought it might benefit later projects.

In the jasmin support json I changed the directory, the file and deleted the
helper. You can now run the jasmine command to test without having to
open the browser.

I also finished the first description. It should be clear.