successfully implemnted menu is hidden test
The menu is hidden test works. I already started coding to test if the
menu hides and display when clicking on the burger icon. Unfortenately I
didn't find out yet how to update the page after an event has happened.
Here I worked with a call back but this is also not good because of
timeOut errors, and it fails to repeatedly test which is necessary.